[WIP] adapt ensamble class
